% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{available_locales}
\alias{available_locales}
\title{Available locales}
\format{
## `available_locales`
A vector with 61 values
}
\usage{
available_locales
}
\description{
A subset of available locales in makeup package
}
\keyword{datasets}
