% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{guess_date_fmt}
\alias{guess_date_fmt}
\title{guess_date_fmt}
\usage{
guess_date_fmt(sample, locale = NULL)
}
\arguments{
\item{sample}{character value with a human date format}

\item{locale}{??}
}
\description{
guesses date format structure for a given sample
}
