% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeup.R
\name{makeup}
\alias{makeup}
\title{makeup}
\usage{
makeup(
  v,
  sample = NULL,
  format = NULL,
  locale = NULL,
  type = NULL,
  suffix = "",
  prefix = ""
)
}
\arguments{
\item{v}{value to be formatted}

\item{sample}{human format to apply in v value}

\item{format}{a character vector of date-time formats. Default is "\%-m/\%-d/\%Y"}

\item{locale}{locale to use, for example "es-MX" for mexican. See posible values at makeup::available_locales}

\item{type}{kind of value to be formatted: ("num", "dat" or "chr")}

\item{suffix}{Character string to append after formatted value}

\item{prefix}{Character string to append before formatted value}
}
\value{
a formatted character value
}
\description{
Formats numbers, strings or dates values for human use
}
\examples{
x <- c(1234.56, 432141, 0.12)

makeup(x, sample = "1'432.1")
makeup(x, sample = "1,432.1")
makeup(x, sample = "10k")
makeup(x, sample = "10\%")

}
