% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{makeup_chr}
\alias{makeup_chr}
\title{makeup_chr}
\usage{
makeup_chr(v, sample = NULL, format = NULL)
}
\arguments{
\item{v}{value to be formatted}

\item{sample}{human format to apply in v value}

\item{format}{??}
}
\value{
a formatted character value
}
\description{
Format numbers for human use
}
\examples{
  makeup_chr("A NAME", sample = "hello")

  makeup_chr("A four word sentence", sample = "Hello Hello")
  makeup_chr(c("A four", " word sentence"), sample = "Hello Hello")



}
