% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{makeup_dat}
\alias{makeup_dat}
\title{makeup_dat}
\usage{
makeup_dat(v, sample = NULL, locale = NULL, format = NULL)
}
\arguments{
\item{v}{value to be formatted}

\item{sample}{human format to apply in v value}

\item{locale}{locale to use, for example "es-MX" for mexican. See posible values at makeup::available_locales}

\item{format}{??}
}
\value{
a formatted character value
}
\description{
Formats dates values for human use
}
\examples{

  v <- as.Date("2020-03-04")
  makeup_dat(v, sample = "Ene 3", locale = "es-CO")

}
