% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeup.R
\name{makeup_format}
\alias{makeup_format}
\title{makeup_format}
\usage{
makeup_format(
  sample = NULL,
  format = NULL,
  locale = NULL,
  type = "",
  suffix = "",
  prefix = ""
)
}
\arguments{
\item{sample}{human format to apply in v value}

\item{format}{??}

\item{locale}{locale to use, for example "es-MX" for mexican. See posible values at makeup::available_locales}

\item{type}{kind of value to be formatted: ("num", "dat" or "chr")}

\item{suffix}{Character string to append after formatted value}

\item{prefix}{Character string to append before formatted value}
}
\value{
a formatted character value

a function
}
\description{
Creates a format sample to be applied to a value
}
\examples{

 f <- makeup_format(sample = "3 de abril 1900")
 f(as.Date("2020-04-28"))
 f <- makeup_format(sample = "abril 3 1900", type = "dat")
 f("2020-04-28")

}
