% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{makeup_num}
\alias{makeup_num}
\title{makeup_num}
\usage{
makeup_num(
  v,
  sample = NULL,
  locale = NULL,
  format = NULL,
  prefix = "",
  suffix = ""
)
}
\arguments{
\item{v}{value to be formatted}

\item{sample}{human format to apply in v value}

\item{locale}{locale to use, for example "es-MX" for mexican. See posible values at available_locales}

\item{format}{??}

\item{prefix}{Character string to append before formatted value}

\item{suffix}{string to append after formatted value}
}
\value{
a formatted character value
}
\description{
Formats numbers values for human use
}
\examples{
 makeup_num(c(0.1, 0.5), "100\%")

}
