% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{which_num_format}
\alias{which_num_format}
\title{which_num_format}
\usage{
which_num_format(str)
}
\arguments{
\item{str}{A number as a character value}
}
\value{
a list with four values: number of decimal values, decimal and thousands separator and specifier value.  The output has the following properties:

$n_decimal

$specifier

$decimal

$thousands
}
\description{
Recognizes decimal values and decimal / thopusands separators. Also adds a specifier
}
\examples{
which_num_format("2135.3")

}
