% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeup.R
\name{makeup_format_js}
\alias{makeup_format_js}
\title{makeup_format_js}
\usage{
makeup_format_js(sample = NULL, locale = NULL, prefix = "", suffix = "")
}
\arguments{
\item{sample}{human format to apply in v value}

\item{locale}{locale to use, for example "es-MX" for mexican. See posible values at makeup::available_locales}

\item{prefix}{Character string to append before formatted value}

\item{suffix}{Character string to append after formatted value}
}
\value{
??
}
\description{
??
}
